Update pull request template to only use checkboxes in the checklist.
Currently the checkboxes also are used to mark alternatives.
Unfortunately in the PR overview, the checkboxes are interpreted as a task do be done and PRs appear to never be finished because not all "tasks" have been checked.

This re-formulates and re-organizes the PR template so that every PR should have all checkbox item checked when finished.