Try to reduce the number of duplicated lines by reducing the notification action to the minimal necessity.
The common interface takes care of the actual notification but to change the graph, access to the "super" object is needed, which is not available in the interface.
A simple callback function is used to execute the code which needs super, but everything else is executed in the default interface implementation.
The resulting code still has duplicated lines (every graph element does essentially the same thing for the labels), but instead of 2 or 3 duplicated lines per function, this only has one duplicated line.