Make the UI integration tests work again.
This means, amongst other changes, to distinguish between the changed container (e.g. a node) and the changed element. Otherwise, e.g. the graph editor can't know when a text overlapping relation was changed and the segments need to be re-calculated. It needs the actual relation to decide this, not only the container (which would be the graph).