Downgrade to Tycho 2.1.0 to avoid problems with our CFF plugin.
The CFF plugin can't reade the dependencies with Tycho 2.2.0 when located in the feature project.
This is probably related to https://bugs.eclipse.org/bugs/show_bug.cgi?id=567760 a new feature to defer dependency resolution.
Since the CFF plugin declares in its metadata that it needs the dependency resolution, I'm not sure if this is a bug in Tycho or our understanding of how Maven works.