Get the affected document graph from the changeset to check if updating the view is necessary.
Previously, when a node/edge was deleted, it was attempted to get the changed graph from the deleted element.
Since it was not attached anymore, the attached graph was none and no update of the view was performed (no need to update the graph editor if the changes concern a different document).

This fixes #326