Update screenshots in graph editor documentation.
Also don't show the whole interface (which could change frequently) in all screenshots, only in the ones where it is important to see the settings or error messages. For the other screenshots, limit them to only show the annotation graph. This hopefully means we don't have to adjust the screenshots that often when the UI changes.