Configure product artifact to execute cff:create and cff:third-party-goal when "cff" plugin is active.
The product holds all the dependency information that is needed. But it is only available, when the root pom is active (otherwise the MANIFEST.MF dependencies are not resolved as Maven dependencies).

But we only want to execute "cff:create" and "cff:third-party-folder" on the product and only when actually needed.

By moving this into a profile, you can execute

mvn -Pcff package

to generate the CITATION.CFF filer and third party folder