Add Font Awesome to citation file and add NOTICE file.
We have a not so common approach to list our third-party dependencies in the CITATION.cff file in the references section.
The NOTICE file is a much more obvious place to look for this list, so add the NOTICE file and explain where to find the needed information. This is more or less the same phrasing as in the "About" dialog.